requirements.txt
fastapi
uvicorn
ics
pydantic
python-dotenv